UserController - registerAdmin, updateUser, getUserTickets ; TicketController - fetch, findById

+++ project/src/main/java/com/todo/project/controller/UserController.java
@@ -2,9 +2,12 @@ package com.todo.project.controller;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.todo.project.exceptions.UserAuthenticationException;
+import com.todo.project.payload.request.UserRequest;
 import com.todo.project.payload.response.UserResponse;
+import com.todo.project.persistence.model.Ticket;
 import com.todo.project.persistence.model.User;
 import com.todo.project.service.HelperService;
+import com.todo.project.service.TicketService;
 import com.todo.project.service.UserService;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
@@ -19,8 +22,11 @@ import java.util.List;
 public class UserController {
     private final UserService userService;
 
-    public UserController(UserService userService){
+    private final TicketService ticketService;
+
+    public UserController(UserService userService, TicketService ticketService){
         this.userService = userService;
+        this.ticketService = ticketService;
     }
 
     @GetMapping("/fetch")
@@ -34,6 +40,7 @@ public class UserController {
             userResponse.setLastName(user.getLastName());
             userResponse.setPassword(user.getPassword());
             userResponse.setEmail(user.getEmail());
+            userResponse.setIsAdmin(user.getIsAdmin());
 
             result.add(userResponse);
         }
@@ -77,6 +84,7 @@ public class UserController {
         }
         try{
             newUser.setPassword(userService.encode(newUser.getPassword()));
+            newUser.setIsAdmin(false);
             userService.createUser(newUser);
             return new ResponseEntity<>("Register successful",HttpStatus.OK);
         }
@@ -85,7 +93,41 @@ public class UserController {
         }
     }
 
+    @PostMapping("register/admin")
+    public ResponseEntity<?> registerAdmin(@RequestBody User newUser, @RequestHeader("session-token") String sessionToken){
+        User admin = userService.findUser(sessionToken);
+        if(admin == null){
+            return new ResponseEntity<>("Incorrect sessionToken", HttpStatus.BAD_REQUEST);
+        }
+        if(!admin.getIsAdmin()){
+            return new ResponseEntity<>("Only admins can create admin accounts", HttpStatus.BAD_REQUEST);
+        }
+        if(!userService.checkUser(newUser).equals("")){
+            return new ResponseEntity<>(userService.checkUser(newUser), HttpStatus.BAD_REQUEST);
+        }
+        if(userService.findUserByEmail(newUser.getEmail()) != null){
+            return new ResponseEntity<>("Email already in use",HttpStatus.BAD_REQUEST);
+        }
+        try{
+            newUser.setPassword(userService.encode(newUser.getPassword()));
+            newUser.setIsAdmin(true);
+            userService.createUser(newUser);
+            return new ResponseEntity<>("Register successful",HttpStatus.OK);
+        }
+        catch (DataIntegrityViolationException e){
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+        }
+    }
 
+    @PutMapping("/update")
+    public ResponseEntity<?> updateUser(@RequestBody UserRequest userRequest){
+        User user = userService.findUserByEmail(userRequest.getEmail());
+        user.setFirstName(userRequest.getFirstName());
+        user.setLastName(userRequest.getLastName());
+        user.setPassword(userRequest.getPassword());
+        userService.updateUser(user);
+        return new ResponseEntity<>("User is saved", HttpStatus.OK);
+    }
 
     @PutMapping("/login")
     public ResponseEntity<?> login(@RequestBody ObjectNode emailAndPasswordInJson){
@@ -95,7 +137,7 @@ public class UserController {
         try{
             user = userService.authenticateAndReturnUser(email, password);
             user.setSessionToken(HelperService.generateNewToken());
-            userService.createUser(user);
+            userService.updateUser(user);
             return new ResponseEntity<>(user, HttpStatus.OK);
         }catch (UserAuthenticationException e){
             return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
@@ -109,18 +151,34 @@ public class UserController {
             return new ResponseEntity<>("Incorrect sessionToken", HttpStatus.BAD_REQUEST);
         }
         user.setSessionToken(null);
-        userService.createUser(user);
+        userService.updateUser(user);
         return new ResponseEntity<>("Logout successful", HttpStatus.OK);
     }
 
     @DeleteMapping("/delete")
-    public ResponseEntity<?> delete(Long id){
+    public ResponseEntity<?> delete(Long id, @RequestHeader("session-token") String sessionToken){
+        User admin = userService.findUser(sessionToken);
+        if(admin == null){
+            return new ResponseEntity<>("Incorrect sessionToken", HttpStatus.BAD_REQUEST);
+        }
+        if(!admin.getIsAdmin()){
+            return new ResponseEntity<>("Only admins can delete accounts", HttpStatus.BAD_REQUEST);
+        }
         User toBeDeleted = userService.findUser(id);
         if(toBeDeleted == null){
             return new ResponseEntity<>("Incorrect id", HttpStatus.NOT_FOUND);
         }
         userService.deleteUser(toBeDeleted);
-        return new ResponseEntity<>("User with id: " + toBeDeleted.getId() + "deleted", HttpStatus.OK);
+        return new ResponseEntity<>("User with id: " + toBeDeleted.getId() + " deleted", HttpStatus.OK);
     }
 
+    @GetMapping("/tickets")
+    public ResponseEntity<?> getUserTickets(@RequestHeader("session-token") String sessionToken){
+        User user = userService.findUser(sessionToken);
+        if(user == null){
+            return new ResponseEntity<>("Incorrect sessionToken.", HttpStatus.BAD_REQUEST);
+        }
+        List<Ticket> tickets = ticketService.findTicketsByCreator(user.getId());
+        return new ResponseEntity<>(tickets, HttpStatus.OK);
+    }
 }
